# packages/contracts/openapi/backend-core/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/users": {
      "get": {
        "operationId": "OrganizationUsersController_listUsers_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List organization users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "OrganizationUsersController_createUser_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create organization user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/sessions": {
      "get": {
        "operationId": "UserController_listUserSessions_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSessionListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List active sessions for a user in tenant scope",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "patch": {
        "operationId": "UserController_patchUser_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "summary": "Patch user in tenant scope",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/sessions:revoke": {
      "post": {
        "operationId": "UserController_revokeUserSessions_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeUserSessionsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Revoke active sessions for a user in tenant scope",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/ai/onboarding:apply": {
      "post": {
        "operationId": "BackendApiController_applyOnboarding_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingApplyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Applied Backend API action response"
          }
        },
        "summary": "Apply a confirmed AI onboarding command (actor_type = ai)",
        "tags": [
          "backend-api"
        ]
      }
    },
    "/api/v1/workflows/backend-api-node:invoke": {
      "post": {
        "operationId": "BackendApiController_invokeNode_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackendApiNodeInvokeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "C5.BackendApiNodeResult"
          }
        },
        "summary": "Apply a change requested by a Workflow node (actor_type = workflow)",
        "tags": [
          "backend-api"
        ]
      }
    }
  },
  "info": {
    "title": "Bridge Backend Core API",
    "description": "M5 C3 Backend REST API: domain CRUD, Communication Core proxy contracts, and stabilized facade endpoints.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000201"
          },
          "organizationId": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000101"
          },
          "telegramUsername": {
            "type": "object",
            "example": "seeded_admin",
            "nullable": true
          },
          "email": {
            "type": "object",
            "example": "seeded-admin@example.bridge.local",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "example": "Seeded Admin"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "blocked"
            ],
            "example": "active"
          },
          "roleCodes": {
            "example": [
              "administrator"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "organizationId",
          "displayName",
          "status",
          "roleCodes",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "Manager User"
          },
          "email": {
            "type": "string",
            "example": "manager@example.bridge.local"
          },
          "telegramUsername": {
            "type": "string",
            "example": "manager_user"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "blocked"
            ],
            "example": "active"
          },
          "roleCodes": {
            "example": [
              "manager"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "displayName"
        ]
      },
      "UserSessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000901"
          },
          "userId": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000201"
          },
          "organizationId": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000101"
          },
          "issuedAt": {
            "type": "string",
            "example": "2026-07-03T10:00:00.000Z"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-07-03T18:00:00.000Z"
          },
          "revokedAt": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "ip": {
            "type": "string",
            "example": "127.0.0.1",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "example": "Mozilla/5.0",
            "nullable": true
          },
          "current": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "userId",
          "organizationId",
          "issuedAt",
          "expiresAt",
          "current"
        ]
      },
      "UserSessionListResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000201"
          },
          "organizationId": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000101"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSessionResponseDto"
            }
          }
        },
        "required": [
          "userId",
          "organizationId",
          "items"
        ]
      },
      "PatchUserDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "Manager User"
          },
          "email": {
            "type": "object",
            "example": "manager@example.bridge.local"
          },
          "telegramUsername": {
            "type": "object",
            "example": "manager_user"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "blocked"
            ],
            "example": "blocked"
          },
          "roleCodes": {
            "example": [
              "administrator",
              "manager"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        }
      },
      "RevokeUserSessionsResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000201"
          },
          "organizationId": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000101"
          },
          "revokedCount": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "userId",
          "organizationId",
          "revokedCount"
        ]
      },
...
      "OnboardingApplyDto": {
        "type": "object",
        "properties": {
          "command": {
            "type": "object",
            "description": "C4 structured command produced by SVC-AI."
          }
        },
        "required": [
          "command"
        ]
      },
      "WorkflowNodeContextDto": {
        "type": "object",
        "properties": {
          "organization_id": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000101"
          },
          "actor_user_id": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000201"
          },
          "trigger": {
            "type": "string",
            "enum": [
              "manual",
              "message",
              "system",
              "workflow"
            ],
            "example": "workflow"
          },
          "roles": {
            "example": [
              "administrator"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "organization_id",
          "actor_user_id",
          "trigger"
        ]
      },
      "BackendApiNodeInvokeDto": {
        "type": "object",
        "properties": {
          "workflow_id": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000401"
          },
          "workflow_version_id": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000402"
          },
          "instance_id": {
            "type": "string",
            "example": "00000000-0000-4000-8000-000000000403"
          },
          "node_id": {
            "type": "string",
            "example": "apply-configuration"
          },
          "context": {
            "$ref": "#/components/schemas/WorkflowNodeContextDto"
          },
          "command": {
            "type": "object",
            "description": "C4 structured command to apply (validated §12.6)."
          }
        },
        "required": [
          "workflow_id",
          "workflow_version_id",
          "instance_id",
          "node_id",
          "context",
          "command"
        ]
      },
...
    }
  },
...
}